+++ src/main/java/edu/grinnell/csc207/texteditor/GapBuffer.java
@@ -12,17 +12,6 @@ public class GapBuffer {
     private int cursor = 0;
     private int size = 0;
 
-
-    public String checkCursor(String message) {
-        if(cursor <= arrLength) { 
-            return ("length is greater than or equal to cursor in ");
-        }
-        if(cursor > 0) { 
-            return ("length is negative in ");
-        }
-        return null;
-    }
-
     /**
      * inserts character to string buffer
      * @param ch - character we want to add
@@ -42,7 +31,6 @@ public class GapBuffer {
         } 
         gapBuffer[cursor] = ch;
         cursor++;
-        checkCursor("insert");
         size++;
         gapStart++;
     }
@@ -56,7 +44,6 @@ public class GapBuffer {
             size--;
             gapStart--;
             cursor--;
-            checkCursor("insert");
         }
     }
 
@@ -88,7 +75,6 @@ public class GapBuffer {
                 gapStart--;
             }
             cursor -= 1;
-            checkCursor("insert");
         }
     }
 
@@ -104,7 +90,6 @@ public class GapBuffer {
                 gapStart++;
             }
             cursor += 1;
-            checkCursor("insert");
         }
     }